PythonXmodemFlash/requirements.txt
pip==25.0.1
pyserial==3.5
xmodem==0.5.0
